requirements.txt
pandas
scikit-learn
numpy
streamlit
requests
python-dotenv
setuptools
wheel